refactor(loading): Obsolete `GetSecretAsync()` and `TryGetSecretAsync()` extension methods.
Some secrets extension methods have been marked as obsolete and will be removed in the future release. Users are encouraged to use the new `Secrets()` store extension method in an attempt to reduce the number of extension functions on the plugin interface as well as streamline access readability in user code.